requirements.txt
pymongo
dnspython==2.2.1
flask==2.2.2
flask_cors==3.0.10
waitress
python-dotenv==0.20.0
argparse==1.4.0
numpy==1.23.2
matplotlib
pandas==1.4.4
